refactor : delete common folder
