+++ utils/survey/convertSurveyComponent.tsx
@@ -2,7 +2,7 @@ import { Chat } from '@/components/custom/Badge/Chat';
 import ButtonWithClickState from '@/components/custom/Button/ButtonWithClickState';
 import ImageItem from '@/components/survey/ImageItem';
 import SurveyItem from '@/components/survey/SurveyItem';
-import { QuestionVO } from '@/types/common/SurveyMessage';
+import { QuestionVO } from '@/types/SurveyMessage';
 
 export function Survey({
   survey,
